+++ src/backend/main/go.backends/index/elasticsearch/messages.go
@@ -74,7 +74,7 @@ func (es *ElasticSearchBackend) SetMessageUnread(user_id, message_id string, sta
 func (es *ElasticSearchBackend) FilterMessages(filter objects.IndexSearch) (messages []*objects.Message, totalFound int64, err error) {
 
 	search := es.Client.Search().Index(filter.User_id.String()).Type(objects.MessageIndexType)
-	search = filter.FilterQuery(search, true).Sort("date_insert", false)
+	search = filter.FilterQuery(search, true).Sort("date_sort", false)
 
 	if filter.Offset > 0 {
 		search = search.From(filter.Offset)