+++ src/backend/main/py.main/caliopen_main/message/objects/message.py
@@ -45,6 +45,7 @@ class Message(ObjectIndexable):
         'date': datetime.datetime,
         'date_delete': datetime.datetime,
         'date_insert': datetime.datetime,
+        'date_sort': datetime.datetime,
         'discussion_id': UUID,
         'external_references': ExternalReferences,
         'identities': [Identity],
@@ -116,7 +117,8 @@ class Message(ObjectIndexable):
         message.is_draft = True
         message.is_received = False
         message.type = "email"  # TODO: type handling inferred from participants
-        message.date_insert = datetime.datetime.now(tz=pytz.utc)
+        message.date = message.date_sort = message.date_insert = \
+            datetime.datetime.now(tz=pytz.utc)
 
         try:
             message.marshall_db()
@@ -202,8 +204,13 @@ class Message(ObjectIndexable):
             current_state["body_plain"] = current_state["body"]
             del (current_state["body"])
 
-        validated_params["current_state"] = current_state
+        # date should reflect last edit time
+        current_state["date"] = self.date
+        current_state["date_sort"] = self.date_sort
+        validated_params["date"] = validated_params["date_sort"] =\
+            datetime.datetime.now(tz=pytz.utc)
 
+        validated_params["current_state"] = current_state
         try:
             self.apply_patch(validated_params, **options)
         except Exception as exc: