+++ src/backend/tools/py.CLI/caliopen_cli/commands/migrate_index.py
@@ -26,7 +26,8 @@ def migrate_index(**kwargs):
         if url and mappings_version:
             client = Elasticsearch(url)
             migration = Migrator(client=client,
-                                 mappings_version=mappings_version)
+                                 mappings_version=mappings_version,
+                                 url=url)
             migration.run()